Exception handles update and delete

+++ src/main/java/com/project/web_servicespring_boot/service/UserService.java
@@ -1,12 +1,14 @@
 package com.project.web_servicespring_boot.service;
 
-import com.project.web_servicespring_boot.entides.User;
+import com.project.web_servicespring_boot.entities.User;
 import com.project.web_servicespring_boot.repositories.UseRepository;
+import com.project.web_servicespring_boot.service.exception.DatabaseException;
 import com.project.web_servicespring_boot.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
-import java.lang.module.ResolutionException;
 import java.util.List;
 import java.util.Optional;
 
@@ -29,13 +31,23 @@ public class UserService {
     }
 
     public void deleteUser(Long id){
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User updateUSer(Long id,User user){
+        try{
         User entity=repository.getReferenceById(id);
         updateData(entity,user);
         return repository.save(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        }
+
     }
 
     private void updateData(User entity, User user) {
